// src/common/Interfaces.d.ts
export interface Location {
  id?: number;
  endOfLine?: boolean;
  fileName?: string;
  logicalLocations?: string[];
  mapped: boolean;
  message?: Message;
  range: Range;
  uri?: Uri;
  uriBase?: string;

  /**
   * Serializes "start" and "stop" properties of VSCode's range as part of the location.
   * That way we can properly type the web view code.
   * @param this Represents the location being serialized.
   * @param key The "key" in the outer object that respresents the location: (i.e. "locationInSarifFile: Location"  - the key is "locationInSarifFile")
   * @param value The current location value.
   */
  toJSON(this: Location, key: any, value: any): any;
}

export interface RunInfo {
  additionalProperties?: { [key: string]: string };
  automationCategory?: string;
  automationIdentifier?: string;
  cmdLine?: string;
  id: number;
  sarifFileFullPath: string;
  sarifFileName: string;
  startUtc?: string;
  timeDuration?: string;
  toolFileName?: string;
  toolFullName?: string;
  toolName: string;

  /**
   * Provides a map between a "baseID" (such as %srcroot%) to its absolute URI.
   * For example a run can contain "originalUriBaseIds" contains an ID of
   * %srcRoot% which had a value of "/src" and a uriBaseId of %driveRoot%
   * "originalUriBaseIds": {
   *    "SRCROOT": {
   *      "uri": "/src",
   *      "uriBaseId": "DRIVEROOT"
   *    },
   *    "DRIVEROOT": {
   *      "uri" : "file:///E:"
   *    },
   * }
   * then this map would contain [ "SRCROOT" : "file:///E:/SRC" , "DRIVEROOT" => "file://E:" ]
   */
  expandedBaseIds?: { [uriBaseId: string]: string };

  workingDir?: string;
}

export interface ResultInfo {
  runInfo: RunInfo;
  additionalProperties?: { [key: string]: string };
  assignedLocation?: Location;
  attachments: Attachment[];
  baselineState: sarif.Result.baselineState;
  codeFlows: CodeFlow[];
  fixes: Fix[];
  id: number;
  kind: sarif.Result.kind;
  locationInSarifFile?: Location;
  locations: Location[];
  message: Message;
  messageHTML?: HTMLLabelElement;
  rank?: number;
  relatedLocs: Location[];
  ruleHelpUri?: string;
  ruleId?: string;
  ruleName?: string;
  ruleDescription?: Message;
  runId: number;
  severityLevel: sarif.Result.level;
  stacks: Stacks;
}

export interface CodeFlow {
  message?: string;
  threads: ThreadFlow[];
}
// ...
export interface Stacks {
  columnsWithContent: StackColumnWithContent;
  stacks: Stack[];
}

export interface Stack {
  frames: Frame[];
  message: Message;
}

export interface Frame {
  location: Location;
  message: Message;
  name: string;
  parameters: string[];
  threadId?: number;
}

export interface Message {
  html?: string;
  text?: string;
}

export interface Attachment {
  description: Message;
  file: Location;
  regionsOfInterest: Location[];
}

export interface Fix {
  description: Message;
  files: FixFile[];
}
// ...
export interface WebviewMessage {
  type: MessageType;
  data: string;
}

export interface DiagnosticData {
  activeTab?: any;
  resultInfo: ResultInfo;
  runInfo: RunInfo;
  selectedRow?: string;
  selectedVerbosity?: any;
}

export interface LocationData {
  eCol: string;
  eLine: string;
  file: string;
  sCol: string;
  sLine: string;
}

export interface ResultsListData {
  columns: { [key: string]: ResultsListColumn };
  filterCaseMatch: boolean;
  filterText: string;
  groupBy?: string;
  groups: ResultsListGroup[];
  resultCount: number;
  sortBy?: ResultsListSortBy;
}

export interface ResultsListGroup {
  rows: ResultsListRow[];
  text: string;
  tooltip?: string;
}

export interface ResultsListRow {
  automationCat: ResultsListStringValue;
  automationId: ResultsListStringValue;
  baselineState: ResultsListBaselineValue;
  kind: ResultsListKindValue;
  logicalLocation: ResultsListStringValue;
  message: ResultsListStringValue;
  rank: ResultsListNumberValue;
  resultFile: ResultsListStringValue;
  resultId: ResultsListNumberValue;
  resultStartPos: ResultsListPositionValue;
  ruleId: ResultsListStringValue;
  ruleName: ResultsListStringValue;
  runId: ResultsListNumberValue;
  sarifFile: ResultsListStringValue;
  severityLevel: ResultsListSeverityValue;
  tool: ResultsListStringValue;
  readonly [key: string]:
    | ResultsListValue
    | ResultsListStringValue
    | ResultsListNumberValue
    | ResultsListCustomOrderValue
    | ResultsListPositionValue;
}
// ...
export interface ResultsListColumn {
  description: string;
  hide: boolean;
  title: string;
}

export interface ResultsListSortBy {
  column: string;
  ascending: boolean;
}